Fix FLOAT type not supported by Postgres

+++ database/models/vehicle.js
@@ -6,7 +6,7 @@ export default class Vehicle extends Model {
     return super.init(
       {
         conso: {
-          type: Sequelize.FLOAT(10, 2),
+          type: Sequelize.NUMERIC,
         },
         createdAt: {
           type: Sequelize.DATE(3),
